exa: factor out 2d/3d compositing
It's a bit difficult to navigate around the code, let's ease the navigation
a tad by splitting the source files.
